feat(energy-facility): refactor out the diesel without litres part
Closes #148
