+++ frontend/src/components/green_house_gaz/energy/HybridMix.vue
@@ -2,18 +2,20 @@
   <div class="content-wrapper">
     <v-col cols="12" sm="6" md="6">
       <v-text-field
-        :value="facility.dieselLiters"
+        :value="localFacility.dieselLiters"
         type="number"
         required
         :rules="rules"
         :min="0"
+        :disabled="localFacility.disableDieselLiters"
         label="Litres of diesel used"
         @input="changeDieselLiters"
       />
     </v-col>
+    <diesel-generator-without-litres :facility.sync="localFacility" />
     <v-col cols="12" sm="6" md="6">
       <v-text-field
-        :value="facility.gridPower"
+        :value="localFacility.gridPower"
         type="number"
         required
         :rules="rules"
@@ -24,7 +26,7 @@
     </v-col>
     <v-col cols="12" sm="6" md="6">
       <v-text-field
-        :value="facility.renewablePower"
+        :value="localFacility.renewablePower"
         type="number"
         required
         :rules="rules"
@@ -36,27 +38,41 @@
   </div>
 </template>
 <script lang="ts">
+import DieselGeneratorWithoutLitres from "@/components/green_house_gaz/energy/DieselGeneratorWithoutLitres.vue";
 import { Facility } from "@/components/green_house_gaz/energy/Facility";
 import { checkRequired, Rule } from "@/utils/rules";
 import { Component, Prop, Vue } from "vue-property-decorator";
 
-@Component
+@Component({
+  components: {
+    DieselGeneratorWithoutLitres,
+  },
+})
 export default class NationalGrid extends Vue {
   @Prop({ type: Object, required: true, default: () => ({}) })
   facility!: Facility;
 
   rules: Rule[] = [checkRequired];
+
+  public get localFacility(): Facility {
+    return this.facility;
+  }
+
+  public set localFacility(value: Facility) {
+    this.$emit("update:facility", value);
+  }
+
   public changeDieselLiters(value: string): void {
     const dieselLiters: number = parseFloat(value || "0");
-    this.$emit("update:facility", { ...this.facility, dieselLiters });
+    this.localFacility = { ...this.facility, dieselLiters };
   }
   public changeGridPower(value: string): void {
     const gridPower: number = parseFloat(value || "0");
-    this.$emit("update:facility", { ...this.facility, gridPower });
+    this.localFacility = { ...this.facility, gridPower };
   }
   public changeRenewablePower(value: string): void {
     const renewablePower: number = parseFloat(value || "0");
-    this.$emit("update:facility", { ...this.facility, renewablePower });
+    this.localFacility = { ...this.facility, renewablePower };
   }
 }
 </script>
